+++ src/main/java/org/example/crossCorrelation/StripesReducer.java
@@ -10,21 +10,21 @@ public class StripesReducer extends Reducer<Text, MapWritable, Text, Text> {
     @Override
     public void reduce(Text key, Iterable<MapWritable> values, Context context)
             throws IOException, InterruptedException {
-        Map<String, Integer> total = new HashMap<>();
+        Map<String, Long> total = new HashMap<>();
         for (MapWritable stripe : values) {
             for (Writable k : stripe.keySet()) {
                 String coItem =  k.toString();
-                IntWritable cnt = (IntWritable) stripe.get(k);
-                total.merge(coItem, cnt.get(), Integer::sum);
+                LongWritable cnt = (LongWritable) stripe.get(k);
+                total.merge(coItem, cnt.get(), Long::sum);
             }
         }
 
         // Сортируем
-        List<Map.Entry<String, Integer>> list = new ArrayList<>(total.entrySet());
-        list.sort(Map.Entry.<String, Integer>comparingByValue().reversed());
+        List<Map.Entry<String, Long>> list = new ArrayList<>(total.entrySet());
+        list.sort(Map.Entry.<String, Long>comparingByValue().reversed());
 
         StringBuilder sb = new StringBuilder();
-        for (Map.Entry<String, Integer> e : list) {
+        for (Map.Entry<String, Long> e : list) {
             if (sb.length() > 0) sb.append(", ");
             sb.append(e.getKey()).append(":").append(e.getValue());
         }
